feat: Draw Cube

+++ bgl/src/d3d12/resource/Rtv_d3d12.cpp
@@ -17,6 +17,6 @@ namespace bgl
 			device->GetDescriptorHandleIncrementSize(descriptorHeap->GetDesc().Type);
 
 		m_CpuHandle = descriptorHeap->GetCPUDescriptorHandleForHeapStart();
-		m_CpuHandle.ptr += descriptorIndex * descriptorSize;
+		m_CpuHandle.ptr += static_cast<size_t>(descriptorIndex) * descriptorSize;
 	}
 }
